bring in multiple external libraries
bring in bandsaw and replace with log4k in order to
make logging more uniform across multiple projects.

bring in config-migrator and make it more similar to
liquibase rather than flyway with a single migration
changelog.

get rid of moshi and replace with config serialization
from spigot-api in order to reduce distributable size.

remove bstats as I don't really refer to it for anything.